fix: Corregir tipo de dato en citas.id_usuario para Foreign Key
ERROR RESUELTO:
- Error 1005 (HY000): Can't create table 'sigc_db'.'citas' (errno: 150)
- Foreign key constraint incorrectly formed

CAUSA:
- citas.id_usuario era INT
- usuarios.id_usuario es BIGINT
- MySQL requiere tipos de datos idénticos para FK

SOLUCIÓN:
- Cambiar citas.id_usuario de INT a BIGINT
- Actualizar migrar_bd.sql para corregir tipo en BD existentes
- Mantener consistencia: usuarios.id_usuario (BIGINT) = citas.id_usuario (BIGINT)

Tipos de datos ahora consistentes:
 usuarios.id_usuario: BIGINT
 citas.id_usuario: BIGINT (corregido)
 doctores.id_doctor: INT
 citas.id_doctor: INT
 horarios.id_horario: INT
 citas.id_horario: INT

+++ crear_bd_completa.sql
@@ -78,7 +78,7 @@ CREATE TABLE citas (
     hora_cita TIME NOT NULL,
     turno VARCHAR(255),
     estado VARCHAR(255) DEFAULT 'ACTIVA',
-    id_usuario INT,
+    id_usuario BIGINT,
     id_doctor INT,
     id_horario INT,
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
